requirements.txt
pydantic~=2.9.2
pandas~=2.2.3
docling~=2.31.0
pyprojroot~=0.3.0
faiss~=1.9.0
numpy~=1.26.4
langchain~=0.3.3
tqdm~=4.66.5
python-dotenv~=1.0.1
openai~=1.51.2
tenacity~=8.5.0
requests~=2.32.3
tabulate~=0.9.0
tiktoken~=0.8.0
aiohttp~=3.10.10
pyyaml~=6.0.2
typer~=0.12.5
pillow~=11.2.1
rich~=14.0.0
torch~=2.0.1+cpu
rtree~=1.4.0
scipy~=1.15.2
beautifulsoup4~=4.13.3
lxml~=5.3.2
marko~=2.1.3
openpyxl~=3.1.5
pypdfium2~=4.30.1
filetype~=1.2.0
click~=8.1.7
setuptools~=75.8.0
pluggy~=1.5.0
pylatexenc~=2.10
pdfplumber~=0.11.6
nltk~=3.9.1
opencv-python~=4.11.0.86
pytesseract~=0.3.13
pdf2image~=1.17.0
click~=8.1.8